+++ eslint.config.mjs
@@ -1,7 +1,6 @@
 import nodePath from 'node:path';
-import {fileURLToPath} from 'node:url';
-import {FlatCompat} from '@eslint/eslintrc';
-
+import { fileURLToPath } from 'node:url';
+import { FlatCompat } from '@eslint/eslintrc';
 import js from '@eslint/js';
 import tseslint from 'typescript-eslint';
 import unicorn from 'eslint-plugin-unicorn';
@@ -9,7 +8,6 @@ import prettierConfig from 'eslint-config-prettier';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = nodePath.dirname(__filename);
-
 const compat = new FlatCompat({
   baseDirectory: __dirname,
 });
@@ -27,17 +25,13 @@ const eslintConfig = [
 
   js.configs.recommended,
   ...tseslint.configs.recommended,
-
   ...compat.extends('next/core-web-vitals'),
-
   unicorn.configs['flat/recommended'],
-
   {
     files: ['**/*.{js,jsx,ts,tsx}'],
     rules: {
       'react/react-in-jsx-scope': 'off',
       'react/prop-types': 'off',
-
       // Unicorn rule override
       'unicorn/filename-case': 'off',
       'unicorn/prevent-abbreviations': 'off',
@@ -52,5 +46,4 @@ const eslintConfig = [
   // IMPORTANT: This MUST be the last configuration in the array.
   prettierConfig,
 ];
-
 export default eslintConfig;
