+++ src/components/ui/chart.tsx
@@ -1,0 +1,385 @@
+"use client"
+
+import * as React from "react"
+import * as RechartsPrimitive from "recharts"
+// Avoid importing complex Recharts prop types to keep compatibility across versions
+
+import { cn } from "@/lib/utils"
+
+// Format: { THEME_NAME: CSS_SELECTOR }
+const THEMES = { light: "", dark: ".dark" } as const
+
+export type ChartConfig = {
+    [k in string]: {
+        label?: React.ReactNode
+        icon?: React.ComponentType
+    } & (
+        | { color?: string; theme?: never }
+        | { color?: never; theme: Record<keyof typeof THEMES, string> }
+    )
+}
+
+type ChartContextProps = {
+    config: ChartConfig
+}
+
+const ChartContext = React.createContext<ChartContextProps | undefined>(undefined)
+
+function useChart() {
+    const context = React.useContext(ChartContext)
+
+    if (!context) {
+        throw new Error("useChart must be used within a <ChartContainer />")
+    }
+
+    return context
+}
+
+function ChartContainer({
+    id,
+    className,
+    children,
+    config,
+    ...props
+}: React.ComponentProps<"div"> & {
+    config: ChartConfig
+    children: React.ComponentProps<
+        typeof RechartsPrimitive.ResponsiveContainer
+    >["children"]
+}) {
+    const uniqueId = React.useId()
+    const chartId = `chart-${id || uniqueId.replaceAll(":", "")}`
+
+    return (
+        <ChartContext.Provider value={{ config }}>
+            <div
+                data-slot="chart"
+                data-chart={chartId}
+                className={cn(
+                    "[&_.recharts-cartesian-axis-tick_text]:fill-muted-foreground [&_.recharts-cartesian-grid_line[stroke='#ccc']]:stroke-border/50 [&_.recharts-curve.recharts-tooltip-cursor]:stroke-border [&_.recharts-polar-grid_[stroke='#ccc']]:stroke-border [&_.recharts-radial-bar-background-sector]:fill-muted [&_.recharts-rectangle.recharts-tooltip-cursor]:fill-muted [&_.recharts-reference-line_[stroke='#ccc']]:stroke-border flex aspect-video justify-center text-xs [&_.recharts-dot[stroke='#fff']]:stroke-transparent [&_.recharts-layer]:outline-hidden [&_.recharts-sector]:outline-hidden [&_.recharts-sector[stroke='#fff']]:stroke-transparent [&_.recharts-surface]:outline-hidden",
+                    className
+                )}
+                {...props}
+            >
+                <ChartStyle id={chartId} config={config} />
+                <RechartsPrimitive.ResponsiveContainer>
+                    {children}
+                </RechartsPrimitive.ResponsiveContainer>
+            </div>
+        </ChartContext.Provider>
+    )
+}
+
+const ChartStyle = ({ id, config }: { id: string; config: ChartConfig }) => {
+    const colorConfig = Object.entries(config).filter(
+        ([, cfg]) => cfg.theme || cfg.color
+    )
+
+    if (colorConfig.length === 0) {
+        return
+    }
+
+    return (
+        <style
+            dangerouslySetInnerHTML={{
+                __html: Object.entries(THEMES)
+                    .map(
+                        ([theme, prefix]) => `
+${prefix} [data-chart=${id}] {
+${colorConfig
+                                .map(([key, itemConfig]) => {
+                                    const color =
+                                        itemConfig.theme?.[theme as keyof typeof itemConfig.theme] ||
+                                        itemConfig.color
+                                    return color ? `  --color-${key}: ${color};` : ""
+                                })
+                                .filter(Boolean)
+                                .join("\n")}
+}
+`
+                    )
+                    .join("\n"),
+            }}
+        />
+    )
+}
+
+const ChartTooltip = RechartsPrimitive.Tooltip
+
+type PayloadItem = {
+    dataKey?: string
+    name?: string
+    value?: number
+    color?: string
+    payload?: Record<string, unknown>
+    fill?: string
+}
+
+type TooltipExtraProps = {
+    payload?: PayloadItem[]
+    label?: unknown
+    active?: boolean
+    hideLabel?: boolean
+    hideIndicator?: boolean
+    indicator?: "line" | "dot" | "dashed"
+    nameKey?: string
+    labelKey?: string
+    labelFormatter?: (value: unknown, payload?: PayloadItem[]) => React.ReactNode
+    labelClassName?: string
+    formatter?: (
+        value: number,
+        name: string,
+        item: PayloadItem,
+        index: number,
+        rawPayload: Record<string, unknown>
+    ) => React.ReactNode
+}
+
+function ChartTooltipContent({
+    active,
+    payload,
+    className,
+    indicator = "dot",
+    hideLabel = false,
+    hideIndicator = false,
+    label,
+    labelFormatter,
+    labelClassName,
+    formatter,
+    color,
+    nameKey,
+    labelKey,
+}: React.ComponentProps<"div"> & TooltipExtraProps) {
+    const { config } = useChart()
+
+    const tooltipLabel = React.useMemo(() => {
+        const items = Array.isArray(payload) ? payload : []
+        if (hideLabel || items.length === 0) {
+            return
+        }
+
+        const [item] = items
+        const key = `${labelKey || item?.dataKey || item?.name || "value"}`
+        const itemConfig = getPayloadConfigFromPayload(config, item, key)
+        const value =
+            !labelKey && typeof label === "string"
+                ? config[label as keyof typeof config]?.label || label
+                : itemConfig?.label
+
+        if (labelFormatter) {
+            return (
+                <div className={cn("font-medium", labelClassName)}>
+                    {labelFormatter(value, payload)}
+                </div>
+            )
+        }
+
+        if (!value) {
+            return
+        }
+
+        return <div className={cn("font-medium", labelClassName)}>{value}</div>
+    }, [
+        label,
+        labelFormatter,
+        payload,
+        hideLabel,
+        labelClassName,
+        config,
+        labelKey,
+    ])
+
+    const items = Array.isArray(payload) ? payload : []
+    if (!active || items.length === 0) {
+        return
+    }
+
+    const nestLabel = items.length === 1 && indicator !== "dot"
+
+    return (
+        <div
+            className={cn(
+                "border-border/50 bg-background grid min-w-[8rem] items-start gap-1.5 rounded-lg border px-2.5 py-1.5 text-xs shadow-xl",
+                className
+            )}
+        >
+            {(!nestLabel) && tooltipLabel}
+            <div className="grid gap-1.5">
+                {items.map((item: PayloadItem, index: number) => {
+                    const key = `${nameKey || item.name || item.dataKey || "value"}`
+                    const itemConfig = getPayloadConfigFromPayload(config, item, key)
+                    const indicatorColor = (item.payload?.fill as string) || item.color || (color as string | undefined)
+
+                    return (
+                        <div
+                            key={String(item.dataKey)}
+                            className={cn(
+                                "[&>svg]:text-muted-foreground flex w-full flex-wrap items-stretch gap-2 [&>svg]:h-2.5 [&>svg]:w-2.5",
+                                indicator === "dot" && "items-center"
+                            )}
+                        >
+                            {formatter && item?.value !== undefined && item.name ? (
+                                formatter(item.value, item.name, item, index, item.payload || {})
+                            ) : (
+                                <>
+                                    {itemConfig?.icon ? (
+                                        <itemConfig.icon />
+                                    ) : (
+                                        !hideIndicator && (
+                                            <div
+                                                className={cn(
+                                                    "shrink-0 rounded-[2px] border-(--color-border) bg-(--color-bg)",
+                                                    {
+                                                        "h-2.5 w-2.5": indicator === "dot",
+                                                        "w-1": indicator === "line",
+                                                        "w-0 border-[1.5px] border-dashed bg-transparent":
+                                                            indicator === "dashed",
+                                                        "my-0.5": nestLabel && indicator === "dashed",
+                                                    }
+                                                )}
+                                                style={
+                                                    {
+                                                        "--color-bg": indicatorColor,
+                                                        "--color-border": indicatorColor,
+                                                    } as React.CSSProperties
+                                                }
+                                            />
+                                        )
+                                    )}
+                                    <div
+                                        className={cn(
+                                            "flex flex-1 justify-between leading-none",
+                                            nestLabel ? "items-end" : "items-center"
+                                        )}
+                                    >
+                                        <div className="grid gap-1.5">
+                                            {nestLabel && tooltipLabel}
+                                            <span className="text-muted-foreground">
+                                                {itemConfig?.label || item.name}
+                                            </span>
+                                        </div>
+                                        {item.value && (
+                                            <span className="text-foreground font-mono font-medium tabular-nums">
+                                                {item.value.toLocaleString()}
+                                            </span>
+                                        )}
+                                    </div>
+                                </>
+                            )}
+                        </div>
+                    )
+                })}
+            </div>
+        </div>
+    )
+}
+
+const ChartLegend = RechartsPrimitive.Legend
+
+type LegendItem = {
+    value?: string
+    dataKey?: string
+    color?: string
+}
+
+type LegendExtraProps = {
+    payload?: LegendItem[]
+    verticalAlign?: "top" | "bottom"
+    nameKey?: string
+    hideIcon?: boolean
+}
+
+function ChartLegendContent({
+    className,
+    hideIcon = false,
+    payload,
+    verticalAlign = "bottom",
+    nameKey,
+}: React.ComponentProps<"div"> & LegendExtraProps) {
+    const { config } = useChart()
+
+    const items = Array.isArray(payload) ? payload : []
+    if (items.length === 0) {
+        return
+    }
+
+    return (
+        <div
+            className={cn(
+                "flex items-center justify-center gap-4",
+                verticalAlign === "top" ? "pb-3" : "pt-3",
+                className
+            )}
+        >
+            {items.map((item: LegendItem) => {
+                const key = `${nameKey || item.dataKey || "value"}`
+                const itemConfig = getPayloadConfigFromPayload(config, item, key)
+
+                return (
+                    <div
+                        key={String(item.value)}
+                        className={cn(
+                            "[&>svg]:text-muted-foreground flex items-center gap-1.5 [&>svg]:h-3 [&>svg]:w-3"
+                        )}
+                    >
+                        {itemConfig?.icon && !hideIcon ? (
+                            <itemConfig.icon />
+                        ) : (
+                            <div
+                                className="h-2 w-2 shrink-0 rounded-[2px]"
+                                style={{
+                                    backgroundColor: item.color as string,
+                                }}
+                            />
+                        )}
+                        {itemConfig?.label}
+                    </div>
+                )
+            })}
+        </div>
+    )
+}
+
+// Helper to extract item config from a payload.
+function getPayloadConfigFromPayload(
+    config: ChartConfig,
+    payload: unknown,
+    key: string
+) {
+    if (!payload || typeof payload !== "object") {
+        return
+    }
+
+    const hasPayloadProp = Object.prototype.hasOwnProperty.call(payload, "payload")
+    const inner = hasPayloadProp ? (payload as Record<string, unknown>).payload : undefined
+
+    const payloadPayload = (inner && typeof inner === "object") ? (inner as Record<string, unknown>) : undefined
+
+    let configLabelKey: string = key
+
+    if (
+        key in (payload as Record<string, unknown>) &&
+        typeof (payload as Record<string, unknown>)[key] === "string"
+    ) {
+        configLabelKey = (payload as Record<string, unknown>)[key] as string
+    } else if (
+        payloadPayload &&
+        key in payloadPayload &&
+        typeof payloadPayload[key] === "string"
+    ) {
+        configLabelKey = payloadPayload[key] as string
+    }
+
+    return configLabelKey in config
+        ? config[configLabelKey]
+        : config[key as keyof typeof config]
+}
+
+export {
+    ChartContainer,
+    ChartTooltip,
+    ChartTooltipContent,
+    ChartLegend,
+    ChartLegendContent,
+    ChartStyle,
+}
